Fix for Centos 5.4 systems where we hit a catch-22 during the mkport.pl build phase, because we need the temporary driver-kmem driver (due to broken kernels and /proc/kcore).

+++ tools/mkdriver.pl
@@ -35,7 +35,7 @@ sub main
 	#   really  need  this  since  all  volatile  #
 	#   stuff is in the build.* dirs anyhow.      #
 	###############################################
-	if ($cmd eq 'clean') {
+	if ($cmd && $cmd eq 'clean') {
 		if (!defined($ENV{BUILD_DIR})) {
 			return system("rm -rf build*");
 		}
